# api/openapi.yaml
...
openapi: 3.0.3
info:
  title: BTP User Containment API
  version: 0.1.0-SNAPSHOT
  description: |
    REST API for the OOBS: a central operator console
    that enrolls SAP BTP subaccounts / IAS tenants, discovers identity providers
    and SoD conflicts, and contains (locks) or unlocks compromised users across subaccounts,
    Cloud Foundry (XSUAA) and IAS in response to an incident.

    ## Authentication
    All `/api/v1/**` endpoints (except `/health` and `/whoami`) require a valid
    XSUAA-issued OAuth2 **bearer JWT**. Authorization is enforced per-operation
    via OAuth2 scopes; the required scope is noted in each operation description.

    ## CSRF
    Keep in mind that the SAP app router uses CSRF. 
    For that see: x-csrf-token docu on the sap help site 
    (it changes all x days so you need to search it)

    ## Scope model
    Four scopes exist, in increasing privilege. The `@PreAuthorize` rules map to:
    - **`btpc.viewer`** - read-only. Satisfied by any of the four scopes.
    - **`btpc.sod_viewer`** - may run SoD scans (admin also allowed).
    - **`btpc.responder`** - may contain/unlock users (admin also allowed).
    - **`btpc.admin`** - full control: enrollment, credentials, config, deletes.

    "Required scope" below names the *minimum* intent; `btpc.admin` always
    satisfies it.

    ## Caller source
    Mutating endpoints accept an optional `X-Caller-Source` header used only for
...
    not change behaviour or authorization.
  license:
...
  contact:
...
servers:
...
security:
  - xsuaa: []
paths:
...
  /api/v1/sod/conflict-sets:
    get:
      tags: [SoD]
      summary: List conflict sets
      description: 'Required scope: `btpc.viewer`.'
      responses:
        '200':
          description: Conflict sets
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ConflictSet' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [SoD]
      summary: Create a conflict set
      description: 'Required scope: `btpc.admin`.'
      parameters:
        - $ref: '#/components/parameters/CallerSource'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ConflictSetPayload' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ConflictSet' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/sod/conflict-sets/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    get:
      tags: [SoD]
      summary: Get one conflict set
      description: 'Required scope: `btpc.viewer`.'
      responses:
        '200':
          description: Conflict set
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ConflictSet' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [SoD]
      summary: Update a conflict set
      description: 'Required scope: `btpc.admin`.'
      parameters:
        - $ref: '#/components/parameters/CallerSource'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ConflictSetPayload' }
      responses:
        '200':
          description: Updated conflict set
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ConflictSet' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [SoD]
      summary: Delete a conflict set
      description: 'Required scope: `btpc.admin`.'
      parameters:
        - $ref: '#/components/parameters/CallerSource'
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/sod/conflict-sets/{id}/enabled:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    patch:
      tags: [SoD]
      summary: Enable or disable a conflict set
      description: 'Required scope: `btpc.admin`.'
      parameters:
        - $ref: '#/components/parameters/CallerSource'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EnabledPatch' }
      responses:
        '200':
          description: Updated conflict set
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ConflictSet' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    xsuaa:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        XSUAA-issued OAuth2 access token. Scopes (`btpc.admin`,
        `btpc.responder`, `btpc.sod_viewer`, `btpc.viewer`) are carried as
        `SCOPE_<name>` authorities and enforced per operation.
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    CallerSource:
      name: X-Caller-Source
      in: header
      required: false
      description: Audit attribution only; default UI.
      schema:
        type: string
        enum: [soar, soar-api, system]
  responses:
    BadRequest:
      description: Invalid input or malformed JSON body
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: Missing or invalid bearer token
    Forbidden:
      description: Authenticated but missing the required scope
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
...
    ErrorResponse:
      type: object
      description: Uniform error envelope. `null` fields are omitted.
      properties:
        error:
          type: string
          description: Stable machine code.
          enum: [bad_request, malformed_body, not_found, conflict, cis_upstream, forbidden, server_error]
        message:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time

    # ---- Platform ----
...
    ConflictSetPayload:
      type: object
      description: >
        ConflictSetService.Payload. Per-kind shape rules apply: `sod` needs ≥2
        role collections, `critical` ≥1, `threshold` needs thresholdCount ≥1,
        `external_email` uses roleCollections as the allowed-domain list.
      required: [name, severity]
      properties:
        name: { type: string }
        description: { type: string, nullable: true }
        severity: { type: string, enum: [low, medium, high, critical] }
        kind: { type: string, enum: [sod, critical, threshold, external_email], default: sod }
        roleCollections:
          type: array
          items: { type: string }
        thresholdCount: { type: integer, nullable: true, description: Only meaningful when kind = threshold. }
        scopeLevel: { type: string, enum: [subaccount, space, org, global], default: subaccount }
    ConflictSet:
      type: object
      description: ninja.gruber.btpc.domain.ConflictSet.
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string, nullable: true }
        severity: { type: string, enum: [low, medium, high, critical] }
        kind: { type: string, enum: [sod, critical, threshold, external_email] }
        roleCollections:
          type: array
          items: { type: string }
        thresholdCount: { type: integer, nullable: true }
        enabled: { type: boolean }
        scopeLevel: { type: string, enum: [subaccount, space, org, global] }
        createdBy: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time, nullable: true }
    EnabledPatch:
      type: object
      required: [enabled]
      properties:
        enabled: { type: boolean }
...